Reduced memory size allocation for attiny to trim off unused bytes in higher memory space.

+++ src/DS2434.cpp
@@ -146,18 +146,18 @@ void DS2434::clearScratchpad(void)
     memset(scratchpad, static_cast<uint8_t>(0xFF), SCRATCHPAD_SIZE);
 }
 
-bool DS2434::writeMemory(const uint8_t* const source, const uint16_t length, const uint16_t position)
+bool DS2434::writeMemory(const uint8_t* const source, const uint8_t length, const uint8_t position)
 {
     if (position >= MEM_SIZE) return false;
-    const uint16_t _length = (position + length >= MEM_SIZE) ? (MEM_SIZE - position) : length;
+    const uint8_t _length = (position + length >= MEM_SIZE) ? (MEM_SIZE - position) : length;
     memcpy(&memory[position],source,_length);
     return true;
 }
 
-bool DS2434::readMemory(uint8_t* const destination, const uint16_t length, const uint16_t position) const
+bool DS2434::readMemory(uint8_t* const destination, const uint8_t length, const uint8_t position) const
 {
     if (position >= MEM_SIZE) return false;
-    const uint16_t _length = (position + length >= MEM_SIZE) ? (MEM_SIZE - position) : length;
+    const uint8_t _length = (position + length >= MEM_SIZE) ? (MEM_SIZE - position) : length;
     memcpy(destination,&memory[position],_length);
     return (_length==length);
 }
